1. Se agregó columna date a la tabla doctor_hours para almacenar la    fecha del día del horario.
2. Se creó vista para gestión de agenda y horarios. Creación,
   actualización y eliminación de horarios por médico.

3. Se estableció manejo de permiso por rol para el acceso al módulo de
   gestión de agenda y horarios.

+++ routes/api.php
@@ -3,8 +3,8 @@
 use App\Http\Controllers\DoctorHoursController;
 use Illuminate\Support\Facades\Route;
 
-Route::get('/doctor/{id}/hours', [DoctorHoursController::class, 'index']);
-Route::post('/doctor-hours', [DoctorHoursController::class, 'store']);
-Route::put('/doctor-hours/{horario}', [DoctorHoursController::class, 'update']);
-Route::delete('/doctor-hours/{horario}', [DoctorHoursController::class, 'destroy']);
-Route::get('/my-hours', [DoctorHoursController::class, 'doctorHours']);
+Route::middleware(['auth', 'authorized:view-agenda'])->get('/doctor/{id}/hours', [DoctorHoursController::class, 'index']);
+Route::middleware(['auth', 'authorized:manage-schedule'])->post('/doctor-hours', [DoctorHoursController::class, 'store']);
+Route::middleware(['auth', 'authorized:manage-schedule'])->put('/doctor-hours/{horario}', [DoctorHoursController::class, 'update']);
+Route::middleware(['auth', 'authorized:delete-schedule'])->delete('/doctor-hours/{horario}', [DoctorHoursController::class, 'destroy']);
+Route::middleware(['auth', 'authorized:view-my-agenda'])->get('/my-hours', [DoctorHoursController::class, 'doctorHours']);
